Refactor se baseando nos inserts com nativequery. Os repositorys serão criados adiante.

+++ src/main/java/dev/friday/com/corretor/repository/LocadorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocadorRepository extends JpaRepository<Locador, Long> {
+public interface LocadorRepository extends JpaRepository<Locador, Integer> {
 }
